+++ swagger_parser/lib/src/config/config_exception.dart
@@ -1,6 +1,9 @@
+/// Exception thrown when a yaml config is not correct
 class ConfigException implements Exception {
+  /// Constructor for [ConfigException]
   const ConfigException(this.message);
 
+  /// A message describing the config error
   final String message;
 
   @override