+++ swagger_parser/lib/src/generator/models/replacement_rule.dart
@@ -1,5 +1,7 @@
+import '../../utils/case_utils.dart';
+
 /// Used to store regex patterns for replacing names during generation
-class ReplacementRule {
+final class ReplacementRule {
   /// Constructor for [ReplacementRule]
   const ReplacementRule({
     required this.pattern,
@@ -12,6 +14,16 @@ class ReplacementRule {
   /// Replacement string
   final String replacement;
 
+  /// Returns the replacement string in the correct case
+  String get replacementInCorrectCase => replacement.isNotEmpty
+      ? replacement[0] == replacement[0].toUpperCase()
+          ? replacement.toPascal
+          : replacement.toCamel
+      : replacement;
+
   /// Applies the replacement rule to the given input string
-  String? apply(String? input) => input?.replaceAll(pattern, replacement);
+  String? apply(String? input) => input?.replaceAll(
+        pattern,
+        replacementInCorrectCase,
+      );
 }