+++ swagger_parser/lib/src/generator/models/universal_rest_client.dart
@@ -3,7 +3,8 @@ import 'package:collection/collection.dart';
 import 'universal_request.dart';
 
 /// Universal template for containing information about Rest client
-class UniversalRestClient {
+final class UniversalRestClient {
+  /// Constructor for [UniversalRestClient]
   const UniversalRestClient({
     required this.name,
     required this.imports,