+++ swagger_parser/lib/src/utils/case_utils.dart
@@ -1,12 +1,15 @@
 /// [CaseUtils] for handling file names and class names
 class CaseUtils {
+  /// Constructor for [CaseUtils]
   CaseUtils(String text) {
     _words = _groupIntoWords(text);
   }
 
   late final List<String> _words;
   static const _separateSymbolsList = r' #,-./@\_{}';
-  static const _upperRegex = r'[A-Z]$';
+  static final _upperCaseRegex = RegExp('[A-Z]');
+  static final _lowerCaseRegex = RegExp('[a-z]');
+  final _upperCaseTwoLettersRowWords = <String>{};
 
   List<String> _groupIntoWords(String text) {
     final sb = StringBuffer();
@@ -15,16 +18,29 @@ class CaseUtils {
 
     for (var i = 0; i < text.length; i++) {
       final char = text[i];
-      final nextChar = i + 1 == text.length ? null : text[i + 1];
       if (_separateSymbolsList.contains(char)) {
         continue;
       }
+
+      final nextChar = i + 1 == text.length ? null : text[i + 1];
+      final nextSecondChar = i + 2 >= text.length ? null : text[i + 2];
+
       sb.write(char);
+
       final isEndOfWord = nextChar == null ||
-          (RegExp(_upperRegex).hasMatch(nextChar) && !isAllCaps) ||
+          (_upperCaseRegex.hasMatch(nextChar) &&
+              !isAllCaps &&
+              (!_upperCaseRegex.hasMatch(char) ||
+                  (nextSecondChar != null &&
+                      _lowerCaseRegex.hasMatch(nextSecondChar)))) ||
           _separateSymbolsList.contains(nextChar);
+
       if (isEndOfWord) {
-        words.add(sb.toString());
+        final word = sb.toString();
+        if (sb.length == 2 && word.toUpperCase() == word) {
+          _upperCaseTwoLettersRowWords.add(word);
+        }
+        words.add(word);
         sb.clear();
       }
     }
@@ -47,14 +63,31 @@ class CaseUtils {
   /// Return text formatted to snake case
   String get snakeCase => _words.map((word) => word.toLowerCase()).join('_');
 
-  String _upperCaseFirstLetter(String word) =>
-      '${word.substring(0, 1).toUpperCase()}${word.substring(1).toLowerCase()}';
+  /// Return text formatted to screaming snake case
+  String get screamingSnakeCase => snakeCase.toUpperCase();
+
+  String _upperCaseFirstLetter(String word) {
+    if (word.length == 2) {
+      final upperCase = word.toUpperCase();
+      if (_upperCaseTwoLettersRowWords.contains(upperCase)) {
+        return upperCase;
+      }
+    }
+
+    return '${word.substring(0, 1).toUpperCase()}${word.substring(1).toLowerCase()}';
+  }
 }
 
 extension StringToCaseX on String {
+  /// Return text formatted to camelCase
   String get toCamel => CaseUtils(this).camelCase;
 
+  /// Return text formatted to PascalCase
   String get toPascal => CaseUtils(this).pascalCase;
 
+  /// Return text formatted to snake_case
   String get toSnake => CaseUtils(this).snakeCase;
+
+  /// Return text formatted to SCREAMING_SNAKE_CASE
+  String get toScreamingSnake => CaseUtils(this).screamingSnakeCase;
 }